Develop

+++ src/unco/features/fuseki.py
@@ -1,16 +1,22 @@
-import os
 import subprocess
 import psutil
 import time
 import requests
+import pandas as pd
+import os
+import signal
+from io import StringIO
+from pathlib import Path
 from unco import UNCO_PATH
+
+OUTPUT_FOLDER  = Path(UNCO_PATH, "data/output")
 
 class FusekiServer:
     """
         Interface to the fuseki server.
     """
 
-    def __init__(self, fuseki_path: str = os.path.join(UNCO_PATH, r"src\fuseki"), gb_ram: int = 4) -> None:
+    def __init__(self, fuseki_path: str = Path(UNCO_PATH, "src/fuseki"), gb_ram: int = 4) -> None:
         """
         Parameters
         ----------
@@ -19,29 +25,46 @@ class FusekiServer:
         mb_ram : int
             RAM size in GB for the fuseki server.
         """
-        self.FUSEKI_PATH = fuseki_path
+        self.FUSEKI_PATH = str(fuseki_path)
         self.GB_RAM = gb_ram
         self.server = None
-        self.starter_path = os.path.join(UNCO_PATH, r"src\unco\features\server_starter.bat")
+        self.starter_path : str
         self._initialize()
 
     def _initialize(self) -> None:
         """
             Method, which creates the server_starter.bat which starts the fuseki server.
         """
-        with open(self.starter_path, 'w') as starter:
-            starter.write("echo 'Running fuseki server'\ncd \"" + self.FUSEKI_PATH + "\"\n")
-            starter.write("java -Xmx" + str(self.GB_RAM) + "G -jar \"fuseki-server.jar\" --update --mem /ds %*")
+        if os.name == "nt":
+            self.starter_path = str(Path(UNCO_PATH, "src/unco/features/server_starter.bat"))
+            with open(self.starter_path, 'w') as starter:
+                starter.write("echo 'Running fuseki server'\ncd \"" + self.FUSEKI_PATH + "\"\n")
+                starter.write("java -Xmx" + str(self.GB_RAM) + "G -jar \"fuseki-server.jar\" --update --mem /ds %*")
+        elif os.name == "posix":
+            self.starter_path = str(Path(UNCO_PATH, "src/unco/features/server_starter.sh"))
+            with open(self.starter_path, 'w') as starter:
+                starter.write(f'FUSEKI_HOME="{self.FUSEKI_PATH}" \nPORT="3030" \ncd "{self.FUSEKI_PATH}" \n./fuseki-server --update --mem /ds &')
+        else:
+            print(f"Unknown system{os.name}. Please contact the admin.")
 
     def start_server(self) -> None:
         """
             Method, which starts the fuseki server.
         """
-        if self.server:
-            raise RuntimeError("Server is already running.")
+        if os.name == "nt":
+            if self.server:
+                raise RuntimeError("Server is already running.")
+            else:
+                self.server = subprocess.Popen(self.starter_path, creationflags=subprocess.CREATE_NEW_CONSOLE, start_new_session=True)
+                time.sleep(3)
+        elif os.name == "posix":
+            if self.server:
+                raise RuntimeError("Server is already running.")
+            else:
+                self.server = subprocess.Popen(['gnome-terminal', '--', 'bash', '-c', 'cd /home/luca/Dokumente/repositories/unco/src/unco/features; ./server_starter.sh; bash'], stdout=subprocess.PIPE, preexec_fn=os.setsid)
+                time.sleep(3)
         else:
-            self.server = subprocess.Popen(self.starter_path, creationflags=subprocess.CREATE_NEW_CONSOLE, start_new_session=True)
-            time.sleep(3)
+            print(f"Unknown system{os.name}. Please contact the admin. Notice: the benchmark is currently not aviable for Mac users.")
     
     def upload_data(self, path: str) -> None:
         """
@@ -60,42 +83,65 @@ class FusekiServer:
             headers = {'Content-Type': r'text/turtle;charset=utf-8'}
         else:
             raise ValueError("Unknown Datatyp. Please use \".rdf\" or \".ttl\" files as input.")
-        data = open(path, 'r', encoding='utf-8').read()
+        data = open(path, 'r', encoding='latin1').read()
         requests.post('http://localhost:3030/ds/data', data=data.encode('utf-8'), headers=headers)
 
 
-    def sparql_query(self, query : str):
+    def delete_graph(self) -> None:
+        """
+            Method, which deletes the current graph ds.
+        """  
+        sparql_query = """
+        DELETE WHERE {
+            ?s ?p ?o
+        }
+        """
+        requests.post(f"http://localhost:3030/ds/update", headers={"Content-Type": "application/sparql-update"}, data=sparql_query)
+
+
+    def run_query(self, query : str, save_result : bool = True) -> pd.DataFrame:
         """
             Method, which runs a given SPARQL query on the fuseki server and outputs the result in json format.
         """
-        headers = {"Accept" : "text/plain"}
-        response = requests.post('http://localhost:3030/ds/sparql', data={'query': query}, headers=headers)
-        return response.text
+        headers = {"Accept" : "text/csv"}
+        params = {"query" : query}
+        response = requests.get('http://localhost:3030/ds/query', params=params, headers=headers)
+
+        csvdata = pd.read_csv(StringIO(response.text), encoding='utf-8')
+
+        if save_result: csvdata.to_csv(str(Path(OUTPUT_FOLDER, "query_results_fuseki.csv")))
+
+        return pd.DataFrame(csvdata)
 
 
     def stop_server(self) -> None:
         """
             Method, which stops the fuseki server.
         """
-        if self.server:
-            for child in psutil.Process(self.server.pid).children():
-                child.kill()
-            self.server = None
+        if os.name == "nt":
+            if self.server:
+                for child in psutil.Process(self.server.pid).children():
+                    child.kill()
+                self.server = None
+            else:
+                raise RuntimeError("Server is already stopped.")
+        elif os.name == "posix":
+            subprocess.call(['pkill', 'gnome-terminal'])
         else:
-            raise RuntimeError("Server is already stopped.")
+            print(f"Unknown system{os.name}. Please contact the admin.")
 
 
 if __name__ == "__main__":
-    f = FusekiServer()
+    f = FusekiServer(Path(UNCO_PATH,"src/apache-jena-fuseki-4.8.0"))
     f.start_server()
-    f.upload_data(r"D:\Dokumente\Repositories\unco\data\output\graph.ttl")
-    # time.sleep(4)
+    f.upload_data(str(Path(UNCO_PATH,"data/output/graph.ttl")))
     query = """
-    PREFIX bsp: <http://beispiel.com/>
+    PREFIX nmo: <http://nomisma.org/ontology#>
+    PREFIX nm: <http://nomisma.org/id/>
 
-    SELECT ?s
+    SELECT ?s ?p ?o
     WHERE {
-        << ?s bsp:fliegtNach ?o >> bsp:dauer 1.4 .
+        <<?s ?p ?o>> ?b1 ?b2 .
     }
     """
     # SELECT ?s { ?s nmo:hasMint nm:comama }
@@ -104,7 +150,8 @@ if __name__ == "__main__":
 
     # SELECT ?s { BIND (<<?s nmo:hasMint nm:comama>> AS ?tripel) ?tripel un:hasUncertainty nm:uncertain_value }
     
-    print(f.sparql_query(query))
+    print(f.run_query(query))
+    time.sleep(7)
     f.stop_server()
